# docs/swagger.yaml
basePath: /api
definitions:
...
  models.Card:
    properties:
      assigned_to:
        type: string
      created_at:
        type: string
      description:
        type: string
      due_date:
        type: string
      id:
        type: string
      list_id:
        type: string
      position:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  models.CreateCardRequest:
    properties:
      assigned_to:
        type: string
      description:
        maxLength: 1000
        type: string
      due_date:
        type: string
      title:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - title
    type: object
...
  models.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  models.UpdateCardRequest:
    properties:
      assigned_to:
        type: string
      description:
        maxLength: 1000
        type: string
      due_date:
        type: string
      list_id:
        type: string
      position:
        type: integer
      title:
        maxLength: 200
        minLength: 1
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API documentation for the Kanban application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Kanban API Documentation
  version: "1.0"
paths:
...
  /organizations/{orgID}/projects/{projectID}/boards/{boardID}/lists/{listID}/cards:
    get:
      description: Retrieves all cards within a specified list. User must own the
        list's board's project's organization.
      parameters:
      - description: Organization ID
        in: path
        name: orgID
        required: true
        type: string
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Board ID
        in: path
        name: boardID
        required: true
        type: string
      - description: List ID
        in: path
        name: listID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of cards
          schema:
            items:
              $ref: '#/definitions/models.Card'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all cards in a list
      tags:
      - Cards
    post:
      consumes:
      - application/json
      description: Creates a new card within a specified list. User must own the list's
        board's project's organization.
      parameters:
      - description: Organization ID
        in: path
        name: orgID
        required: true
        type: string
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Board ID
        in: path
        name: boardID
        required: true
        type: string
      - description: List ID
        in: path
        name: listID
        required: true
        type: string
      - description: Card creation details
        in: body
        name: card
        required: true
        schema:
          $ref: '#/definitions/models.CreateCardRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Card created successfully
          schema:
            $ref: '#/definitions/models.Card'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new card
      tags:
      - Cards
  /organizations/{orgID}/projects/{projectID}/boards/{boardID}/lists/{listID}/cards/{cardID}:
    delete:
      description: Deletes a specific card by its ID within a specified list. User
        must own the list's board's project's organization.
      parameters:
      - description: Organization ID
        in: path
        name: orgID
        required: true
        type: string
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Board ID
        in: path
        name: boardID
        required: true
        type: string
      - description: List ID
        in: path
        name: listID
        required: true
        type: string
      - description: Card ID
        in: path
        name: cardID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a card
      tags:
      - Cards
    get:
      description: Retrieves a specific card by its ID within a specified list. User
        must own the list's board's project's organization.
      parameters:
      - description: Organization ID
        in: path
        name: orgID
        required: true
        type: string
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Board ID
        in: path
        name: boardID
        required: true
        type: string
      - description: List ID
        in: path
        name: listID
        required: true
        type: string
      - description: Card ID
        in: path
        name: cardID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Card details
          schema:
            $ref: '#/definitions/models.Card'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get card by ID
      tags:
      - Cards
    put:
      consumes:
      - application/json
      description: Updates a specific card by its ID within a specified list. User
        must own the list's board's project's organization. Supports moving card to
        another list.
      parameters:
      - description: Organization ID
        in: path
        name: orgID
        required: true
        type: string
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Board ID
        in: path
        name: boardID
        required: true
        type: string
      - description: List ID
        in: path
        name: listID
        required: true
        type: string
      - description: Card ID
        in: path
        name: cardID
        required: true
        type: string
      - description: Card update details
        in: body
        name: card
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Card updated successfully
          schema:
            $ref: '#/definitions/models.Card'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update a card
      tags:
      - Cards
...
securityDefinitions:
  ApiKeyAuth:
    description: Enter your JWT token in the format "Bearer <your_token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...